move omakase modal outside omakase folder, update to cloudscape styles and implement alerting
