--- frontend/src/views/OmakasePlayer/components/OmakaseModal/OmakaseExportModal.css
@@ -1,284 +1,0 @@
-.segmentation-export {
-  padding: 0.5em 1em;
-  display: flex;
-  align-items: center;
-  gap: 5px;
-  cursor: pointer;
-}
-
-.segmentation-export-disabled {
-  padding: 0.5em 1em;
-  display: flex;
-  align-items: center;
-  gap: 5px;
-  cursor: default;
-  opacity: 0.3;
-}
-
-.omakase-select {
-  width: 100%;
-  background-color: #454853;
-  border: none;
-  border-bottom: 1px solid #ffffff;
-  border-radius: 0px;
-  padding-left: 0;
-  padding-right: 2.5em;
-
-  color: #ffffff;
-  font-size: 13px;
-  line-height: 18px;
-
-  padding-bottom: 0.2em;
-
-  appearance: none;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  position: relative;
-  cursor: pointer;
-  outline: none;
-}
-
-.omakase-select option {
-  width: 100%;
-}
-
-label {
-  font-size: 13px;
-  line-height: 26px;
-  font-weight: 400;
-}
-
-.omakase-select-wrapper {
-  position: relative;
-  display: inline-block;
-  margin-bottom: 1em;
-}
-
-.omakase-select-wrapper select {
-  width: 100%;
-}
-.omakase-select:focus {
-  outline: none;
-  box-shadow: none;
-}
-
-select {
-  box-shadow: none;
-  outline: none;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-}
-
-.omakase-select-wrapper svg {
-  position: absolute;
-  right: 10px;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 16px;
-  height: 16px;
-  pointer-events: none;
-}
-
-.omakase-checkbox input {
-  position: absolute;
-  opacity: 0;
-  cursor: pointer;
-  width: 0;
-  height: 0;
-}
-
-.omakase-checkbox {
-  display: flex;
-  align-items: center;
-  font-size: 13px;
-  color: #ffffff;
-  gap: 10px;
-  margin-top: 1em;
-}
-
-.checkmark {
-  width: 20px;
-  height: 20px;
-  background-color: #454853;
-  border: 2px solid #ffffff;
-  display: inline-block;
-  border-radius: 4px;
-  position: relative;
-}
-
-.omakase-checkbox input:checked + .checkmark::after {
-  content: "";
-  position: absolute;
-  left: 5px;
-  top: 2px;
-  width: 6px;
-  height: 10px;
-  border: solid white;
-  border-width: 0 2px 2px 0;
-  transform: rotate(45deg);
-}
-
-.omakase-checkbox-icon {
-  display: flex;
-  align-items: center;
-
-  height: 18px;
-}
-
-.omakase-checkbox-icon-container {
-  height: 18px;
-  cursor: pointer;
-}
-
-.omakase-dropdown {
-  background-color: #323544;
-  border-radius: 5px;
-  margin-top: 2em;
-  padding: 0em 2em 1em 2em;
-}
-
-.omakase-dropdown-title {
-  font-size: 13px;
-}
-
-.omakase-dropdown-icon {
-  height: 18px;
-}
-
-.omakase-dropdown-toggle-control {
-  padding-top: 1em;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  cursor: pointer;
-  color: #ffffff;
-}
-
-.omakase-dropdown-toggle-control-disabled {
-  padding-top: 1em;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  color: #cacfea;
-  opacity: 0.4;
-}
-
-.omakase-input {
-  /* remove default styling */
-  border: none;
-  background: none;
-  outline: none;
-  box-shadow: none;
-  padding: 0;
-  margin: 0;
-  appearance: none;
-
-  width: 100%;
-  border-bottom: 1px solid #ffffff;
-
-  font-size: 13px;
-  color: #ffffff;
-
-  margin-bottom: 1.5em;
-}
-
-.omakase-input:first-child {
-  /* remove default styling */
-  margin-top: 1em;
-}
-
-input::placeholder {
-  color: #cacfea;
-}
-
-.omakase-input-short {
-  width: 200px;
-}
-
-.omakase-export-button {
-  background: none;
-  border: none;
-  padding: 0;
-  margin: 0;
-  font: inherit;
-  color: inherit;
-  text-align: inherit;
-  appearance: none;
-  cursor: pointer;
-
-  width: 120px;
-  height: 40px;
-
-  background-color: #662d91;
-  border: 1px solid #9159ba;
-  border-radius: 5px;
-
-  color: #ffffff;
-}
-
-.omakase-export-button:disabled {
-  opacity: 0.4;
-  cursor: default;
-}
-
-.omakase-cancel-button {
-  background: none;
-  border: none;
-  padding: 0;
-  margin: 0;
-  font: inherit;
-  color: inherit;
-  text-align: inherit;
-  appearance: none;
-  cursor: pointer;
-
-  width: 120px;
-  height: 40px;
-
-  background-color: #4c5578b2;
-  border: 1px solid #7c8bc2;
-  border-radius: 5px;
-
-  color: #ffffff;
-}
-
-.omakase-button-container {
-  display: flex;
-  align-items: center;
-  justify-content: end;
-
-  margin-top: 3em;
-  gap: 1em;
-}
-
-.omakase-button-spinner-container {
-  display: flex;
-  align-items: center;
-  gap: 5px;
-}
-
-.omakase-toast {
-  position: fixed;
-  bottom: 16px;
-  right: 16px;
-  background-color: #333;
-  color: white;
-  padding: 12px 16px;
-  text-transform: none;
-  border-radius: 4px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  opacity: 1;
-  transition: opacity 0.3s ease-in-out;
-  z-index: 1000;
-}
-
-.omakase-toast-success {
-  background-color: #00b69a;
-  border: 1px solid #00e9a3;
-}
-
-.omakase-toast-error {
-  background-color: #ea5a5a;
-  border: 1px solid #ff8282;
-}
