--- frontend/src/views/OmakasePlayer/components/OmakaseModal/OmakaseModal.css
@@ -1,67 +1,0 @@
-.modal-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    background-color: rgba(0, 0, 0, 0.6);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-    backdrop-filter: blur(1px);
-  }
-  
-  .modal-content {
-    background-color: #454853;
-    padding: 24px;
-    border-radius: 4px;
-    border: 1px solid #656b7e;
-    position: relative;
-    width: 450px;
-    max-width: 90%;
-    box-shadow: 0 8px 16px rgba(0, 0, 0, 0.3);
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .modal-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    font-size: 15px;
-    font-weight: 700;
-    line-height: 18px;
-    color: #CACFEA;
-    padding-bottom: 12px;
-    
-  }
-
-  .modal-header-only-close {
-    justify-content: end;
-  }
-  
-  .modal-title {
-    flex-grow: 1;
-    text-align: left;
-  }
-  
-  .modal-close {
-    background: none;
-    border: none;
-    font-size: 18px;
-    color: #aeb3c1;
-    cursor: pointer;
-    transition: color 0.2s ease-in-out;
-  }
-  
-  .modal-close:hover {
-    color: #ffffff;
-  }
-  
-  .modal-body {
-    margin-top: 16px;
-    font-size: 16px;
-    color: #d1d5db;
-  }
-
